Type fixes, 64-bit build and AltirraRMT 4.10 DLLs x86 and x64 builds

+++ cpp_src/IO_Song.cpp
@@ -1079,7 +1079,7 @@ bool CSong::LoadTxt(std::ifstream& in)
 					Trimstr(value);
 					memset(m_songname, ' ', SONG_NAME_MAX_LEN);
 					int lname = SONG_NAME_MAX_LEN;
-					if (strlen(value) <= SONG_NAME_MAX_LEN) lname = strlen(value);
+					if (strlen(value) <= SONG_NAME_MAX_LEN) lname = (int) strlen(value);
 					strncpy(m_songname, value, lname);
 				}
 				else
